repo/pyproject.toml
[project]
name = "offline-platform-api"
version = "0.1.0"
description = "Offline Retail Checkout and Entrepreneurship Project Incubation API"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "alembic>=1.16.0",
  "cryptography>=45.0.0",
  "fastapi>=0.116.0",
  "passlib[argon2]>=1.7.4",
  "PyJWT>=2.9.0",
  "psycopg[binary]>=3.2.0",
  "pydantic-settings>=2.10.0",
  "python-multipart>=0.0.20",
  "sqlalchemy>=2.0.40",
  "uvicorn[standard]>=0.35.0",
]

[project.optional-dependencies]
dev = [
  "httpx>=0.28.0",
  "pytest>=8.3.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = ["app"]
